refactor: merge FileUploadConfig (#22)
